Convert Lexer class from Java to Scala
using javatoscala.com
The code in line 44 had to be manually fixed:
```
acumulator.append(tokenArray(tokenPos)) { tokenPos += 1; tokenPos - 1 }
```
=>
```
acumulator.append(tokenArray(tokenPos))
tokenPos += 1
```
The code converter is not smart enough to distinguish between
tokenPos++ as a statement that's just there for the side-effect
and tokenPos++ as an expression where gets the current values of tokenPos before
it is increased by 1.

The same has been done (not absolutely necessarily, but simplifies the code) in
other places.